Feature/python cleanup

+++ Source/cielim-python/cielim/driver.py
@@ -1,45 +1,79 @@
-import struct
+import random
 
 import cv2
 import delimited_protobuf
 import numpy as np
 import zmq
-import random
 
-import cielimMessage_pb2 as cielimMessage
-import imageDiagnostics_pb2 as imageDiagnostics
+from .cielimMessage_pb2 import CielimMessage
+from .imageDiagnostics_pb2 import DiagnosticData
+from .scene import Scene
 
 
 class Connector:
+    """
+    A class to manage connections to Cielim and send/receive messages.
+    """
+
     def __init__(self):
         self.address = ""
         self.context = None
         self.request_socket = None
         self.identity = ""
 
-    def connect(self, address: str = "tcp://localhost:5556"):
+    def connect(self, address: str = "tcp://localhost:5556", socket_timeout_ms: int = 10000):
         # Ensure each connector has different ID
         self.identity = "CielimConnector" + "".join(random.choices("0123456789", k=5))
 
         self.context = zmq.Context()
         self.request_socket = self.context.socket(zmq.REQ)
         self.request_socket.set(zmq.SocketOption.CONNECT_TIMEOUT, 10)
+        # self.request_socket.set(zmq.SocketOption.RCVTIMEO, socket_timeout_ms)
         self.request_socket.setsockopt_string(zmq.IDENTITY, self.identity)
 
         self.address = address
         self.request_socket.connect(address)
-        print(self._send_ping())
-        print(self._send_ping())
 
-    """
-    Similar to normal blocking recv_multipart but if message received is a
-    PING (indicating a heartbeat request), we respond with PONG immediately
-    and then continue waiting until data is received.
-    """
+        # Send a ping to ensure connection is established before sending any data
+        print(f"Sending ping to Cielim at {address}...")
+        print(self.send_ping())
 
-    def safe_recv_multipart(self):
+        # Send initial dummy scene with white sphere, otherwise first image will be blank
+        # TODO: Find a way to fix this Cielim-side
+
+        dummy_scene = Scene()
+
+        # 180 degree rotation about the y-axis, facing towards the origin
+        dummy_scene.set_spacecraft_params(position=(0, 0, 2000), attitude=(0, 1, 0))
+
+        index = dummy_scene.add_celestial_body("asteroid1")
+        dummy_scene.set_celestial_body_params(
+            index, mesh_shape="sphere_normalized", mesh_brdf="Lambertian", mesh_radius=1000
+        )
+
+        index = dummy_scene.add_celestial_body("asteroid2")
+        dummy_scene.set_celestial_body_params(
+            index, position=(0, 0, 200), mesh_shape="sphere_normalized", mesh_brdf="Regolith", mesh_radius=100
+        )
+
+        self.send_init_request()
+        self.send_frame(dummy_scene.get_scene())
+        self.request_image_for_camera_id(1, True, False)
+
+        self.send_init_request()  # Reset scene
+
+    def _safe_recv_multipart(self):
+        """
+        Similar to normal blocking recv_multipart but if message received is a PING (indicating a heartbeat request), we respond with PONG immediately and then continue waiting until data is received.
+        """
+        if self.request_socket is None:
+            raise Exception("Socket not connected, cannot receive messages.")
+
         while True:
-            multipart_message = self.request_socket.recv_multipart()
+            try:
+                multipart_message = self.request_socket.recv_multipart()
+            except zmq.error.Again:
+                raise Exception("Socket timed out waiting for message.")
 
             if multipart_message[0] == b"PING":
                 self.request_socket.send_string("PONG")
@@ -47,17 +81,29 @@ class Connector:
 
             return multipart_message
 
-    def _send_ping(self):
+    def send_ping(self):
+        if self.request_socket is None:
+            raise Exception("Socket not connected, cannot send ping.")
+
         _ = self.request_socket.send_string("PING")
-        return self.safe_recv_multipart()[0].decode("utf-8")
+
+        return self._safe_recv_multipart()[0].decode("utf-8")
 
     def send_init_request(self):
-        _ = self.request_socket.send_string("INIT_SCENE")
-        return self.safe_recv_multipart()[0].decode("utf-8")
+        if self.request_socket is None:
+            raise Exception("Socket not connected, cannot send init request.")
 
-    def send_frame(self, sim_frame: cielimMessage.CielimMessage):
+        _ = self.request_socket.send_string("INIT_SCENE")
+
+        return self._safe_recv_multipart()[0].decode("utf-8")
+
+    def send_frame(self, sim_frame: CielimMessage):
+        if self.request_socket is None:
+            raise Exception("Socket not connected, cannot send frame.")
+
         _ = self.request_socket.send_multipart([b"SIM_UPDATE", sim_frame.SerializePartialToString()])
-        return self.safe_recv_multipart()[0].decode("utf-8")
+
+        return self._safe_recv_multipart()[0].decode("utf-8")
 
     def request_image_for_camera_id(
         self,
@@ -66,6 +112,9 @@ class Connector:
         should_return_diagnostics: bool = True,
         format_raw: bool = False,
     ):
+        if self.request_socket is None:
+            raise Exception("Socket not connected, cannot request image.")
+
         _ = self.request_socket.send_multipart(
             [
                 b"REQUEST_IMAGE",
@@ -75,7 +124,7 @@ class Connector:
             ]
         )
 
-        [image_data, image_data_size, diagnostics_serialized] = self.safe_recv_multipart()
+        [image_data, image_data_size, diagnostics_serialized] = self._safe_recv_multipart()
 
         image = None
         if should_return_image:
@@ -86,7 +135,7 @@ class Connector:
             else:
                 image = cv2.imdecode(buf, cv2.IMREAD_COLOR)  # Format PNG image data for viewing with OpenCV (BGRA)
 
-        diagnostics = imageDiagnostics.DiagnosticData()
+        diagnostics = DiagnosticData()
         diagnostics.ParseFromString(diagnostics_serialized)
 
         cob = (diagnostics.cob_x, diagnostics.cob_y)
@@ -96,38 +145,46 @@ class Connector:
         return [image, cob, coverage]
 
     def disconnect(self):
-        self.request_socket.disconnect(self.address)
-        self.context.destroy()
+        if self.request_socket is not None:
+            self.request_socket.disconnect(self.address)
+
+        if self.context is not None:
+            self.context.destroy()
 
 
-def create_simulation_frame():
-    return protobuf_data
+class CielimMessageFileHandler:
+    """
+    A class to handle CielimMessage bin files and reading.
+    """
 
-
-class MessageFileHandler:
     def __init__(self, file_name: str):
         self.file_name = file_name
         self.file_handle = open(file_name, "rb")
 
-    def _read_simulation_frame_at_time(self, sim_time: float):
-        proceed = True
-        while proceed:
-            message = delimited_protobuf.read(self.file_handle, cielimMessage.CielimMessage)
-            if cielimMessage.currentTime.simTimeElapsed >= sim_time:
-                proceed = False
+    def jump_to_simulation_frame_at_time(self, sim_time: float):
+        """
+        Jumps to first message in file at or after the requested sim time and returns that message.
+        """
+        while True:
+            message = delimited_protobuf.read(self.file_handle, CielimMessage)
+
+            if message is None:
+                raise Exception("Reached end of file without finding a message with sim time >= requested time.")
+
+            if message.currentTime.simTimeElapsed >= sim_time:
                 return message
 
     def get_simulation_frame_at_time(self, sim_time: float):
-        message = self._read_simulation_frame_at_time(sim_time)
-        self.file_handle.seek(0)
+        """
+        Get first message in file at or after the requested sim time and reset file reader to start of file.
+        """
+        message = self.jump_to_simulation_frame_at_time(sim_time)
+        self.file_handle.seek(0)  # Reset file reader to beginning of file after finding requested message
         return message
-
-    def jump_to_simulation_frame_at_time(self, sim_time: float):
-        return self._read_simulation_frame_at_time(sim_time)
 
     def get_next_simulation_frame(self):
         try:
-            return delimited_protobuf.read(self.file_handle, cielimMessage.CielimMessage)
-        except:
-            print("Protobuf failed to decode.")
+            return delimited_protobuf.read(self.file_handle, CielimMessage)
+        except Exception as e:
+            print(f"Protobuf failed to decode: {e}")
             return None
